backend/requirements.txt
# ===============================
# Core runtime (Python 3.10+)
# ===============================

# YAML parsing (mkdocs.yml, frontmatter)
PyYAML>=6.0.3

# Markdown parsing (AST-level, reliable)
markdown-it-py>=4.0.0
mdit-py-plugins>=0.5.0

# ===============================
# Data processing
# ===============================

pandas>=2.3.3

# ===============================
# PDF ingestion
# ===============================

PyPDF2>=3.0.1

# ===============================
# Environment variables
# ===============================

python-dotenv>=1.2.1

# ===============================
# Google Gemini SDK
# ===============================

google-genai>=1.60.0

# ===============================
# Fast Task API for async tasks
# ===============================
fast-task-api>=1.3.3